Add/alter some batch peak fit options to accomidate batch GUI.
Added option to provide already-parsed spectrum files to `fit_peaks_in_files(...)` (this is so spooled files can be used, but thier oriinal names will appear in the reports).
Optionally return most of the results in-memorry, instead of just output files.
Added an option to not need an exemplar input file, but instead just fit for all peaks.