Move back to only requiring C++17.
Using C++20 is probably doable, but came with some hassle.
Wt 3.7.1 can not be compiled under C++20, but it would probably be fine to just use C++17 for it, and C++20 for everything else.
On RHEL using gcc 13, Wt::Signal specialization for void needed to be patched so c++20 code that included it would compile; not that big of a deal.
On Windows also ran into some difficulties getting all the prerequisits to compile, that could probably be overcome, but was just taking too much time.
And on macOS it required moving to a OSX target of 13.3, which could be fine, but would leave some computers behind.
So for now, will just use C++17.