Fix uploading multiple files on macOS/electron from drag-n-drop, and a few other small batch-gui related things.
There was an exception in JS, due to some bad code; now fixed in both relevant places hopefully.
Start preloading batch GUI CSS so its avaiable before the widget is initially shown, so layout will be correct from the begining.
Maybe fix a compile issue in PeakFitLM having to do with isnan and isinf on the different platforms.