Fix "Min BR" not correctly working on the "Nuclide Search" tab.
A user reported bug that the user enterable minimum BR field did not work as expected, for non-zero values.
Turns out was using using fraction of decay gammas, not the intended relative BR, relative to max gamma.
Also, was not consistently using intensity relative to max intensity gamma of the decay chain.
I think fixed all this up, so it behaves as the user expects, although possibly at the cost of a little more CPU time.
Also added unit test to spot check a few cases.