For RelActAuto, add using rough initial activity guess method for rel. act. that are zero from manual solution.
This is to help get rel. act. from starting at zero, if even the manual solution said it should be.  Its by no means great, and the rough estimation (and when to use it) could use more work, but a step in the right direction.

Also added button to sorti the displayed energy ranges.
A few other little fixes.