Sketch out improving QR code searching - possible non-working code.
Move to always using a `canvas` element to first display the image, and then converting to PNG.
Add in a number of adjustments to the original image, to try and help zxing to search for barcode.
JS was changed based on stand-alone test implementation, but not tested once integrated into InterSpec.