First go at making skeletons/proof-of-concepts for LLM MCP Server REST interface, as well as direct interface to use LLM.
This is more of explore what is needed implementation, for future planning.  Seems to communicate with a locally hosted LLM on just a dummey request to trigger a tool call. Looks like we can avoid using a 3rd party library to facilitate LLM communications.

Things very much are in the sketch-it-out phase, and not totally working, and not architected very well; things will change, if continued to be pursued.