Add skeleton LLM GUI, implement some more tool calls, fix a few things.
Added a very basic GUI to allow interacting with the LLM; it opens a new tool tab, and its presence is saved as part of the app state.
Added chat history to N42 saving (not yet verified to be functional)
Fixed some issues with saving/reconstructing the chat history for submitting new questions.
Added a couple more tool calls, for fitting peaks, and retrieving user peaks.
Added some basic capibilities to set peaks to background/secondary spectrum; this capability will most likely be expanded in the future to enable some user work-flows.