Continue cleaning up implementation ever so slightly, and add a couple more tool calls
Add in tool call to fit peaks for nuclide(s).  This is just a very rough-in; this needs work to be more reliable and do a better job - but kinda works for now.
Add being able to get automated RIID results bia tool calls.
Also fixed potential bug in RelActAuto if two user-peaks had identical energies.