Fix issue with batch peak fitting failing when the live-time is zero, due to errors templating the reults.
Setting a JSON file to inf/nan would set the actualy JSON value to null (since inf/nan are not allowed in JSON), which would cause the c++ `printFixed(...)` function to throw an exception when templating results.  Resolved by checking for argument values, and just printing a "--" in this case.  Also applied some slight more hardening to this function and its `printCompact(...)` sibling.