Added non-background subtraction peak equivalents for RelActAuto.
When the RelActAuto solution is fit with background subtraction, the fit peaks cant be directly plotted on the non-background subtracted foreground, as the continuum (and hence whole peak) will be low, by the background amount.  So added a set of peaks, `solution.m_peaks_without_back_sub`, wher ethe continuums, but no other quantities of the fit peaks have been re-fit.  I *think* this should be roughly, kinda, okay, for plotting purposes - testing is needed.