Added offset for activity paramters for RelActAuto.
Behaviour of Ceres paramaters with a value of 0.0 doesnt seem to be very good, and it looks like once a value of 0.0 was tried for a activity parameter, it often times would never move to a different value.  To hopefully avoid this, gave the activity parameters a offset of 1.0, so they never get to zero.  Seems to compute okay, but havent tested on spectra where the issue seemed to popup.

Also, fixed up some bugs and bad asserts for activity ratio constraints.