Fix issue where RelActAuto activities would get stuck at zero.
When computing contributions of peaks, the code was happily skipping zero-amplitude peaks - however, since the Jacobians of the residuals are being computed using auto-differentiation, the gradient of the solution, wrt the activities would become flat if a sources activity was ever tried at zero, and the activity would perhaps get stuck there.  Changed it so controbution of zero-amplitude peaks are now accounted for in peaks.
Also fixed computation of rel act uncertainities.
Added rel. act. printout to textt results.
And arbitrarily adjust initial activity estimate, if the RelActManual estimate has a large uncertainty; this may not be needed now that the amplitude==0 thing is fixed, but was seeing some unreasonable initial activities, for sources with large RelActManual activity uncertainities, so leaving in.