Get kinetic reference lines seemingly working, and roughed in which nuclides are avaiblable, but still need much more work and care on this.
The Kinetic reference lines seem to work decently well; the client-side avaiable nuclides are based off external RID, on-board-RID, IDs associated with peaks already, and the "always" lines specified in data/kinetic_ref_lines.xml.
Roughed in suggesting additional nuclides, based on peaks in the spectrum, but further filtering of those by picking out the best ones are necassary to not overwhelm the client side.
Still need to filter the lines sent to the client so there arent so many (both for transfer size, and latency when moveing the mouse).
Also change auto peak search to be multithreaded.