Work on kinetic reference lines.
Added a default color for kinetic reference lines to the ColorTheme, and to the GUI that allows customizing it.
Added in re-generating refernce lines when color them changes.
Refactored `profile_weight(...)` a bit so we could use it to judge a sources compatibility with the spectrum.
Add some more commone nuclides to the "always" shor reference lines.
Added kinetic reference lines for:
- escape peaks
- random sum peaks
- based on nuclide suggestions/characteristic lines
- associated nuclides.

Currently the computation of these kinetic reference lines takes far too long to leave inside the event loop, so will need to compute in a background thread.