Move part of the prepairing spectrum file to send to external RID service, to not be in the application thread.
The writing of data in `RestRidInputResource::handleRequest(...)` to a N42 file could be a little slow, so move that work to not be within the WApplication::UpdateLock.
This was fine to do because `RemoteRid::fileForAnalysis(...)` returns a bespokely constructed SpecFile object.
Also, switch to using `Wt::Json` to create the options JSON - jic the DRF name needs escaping or something.