Add using Berstein polynomials for FWHM computation for RelActAuto.
In principle this should be more stable, and allow efficiently/nicely set the lower/upper limits of FWHM using Ceres Solver paramater limits.
To add in Berstein, we also have to track the lower/upper energies we expect to call FWHM for - tried to rough this in - but may need a little more tweaking.
Seems to work for simple cases - but needs more debugging.