Try to get RelActAuto to converge on correct solution a little better.
Changed default detector efficeincy to a generic coax, since it might be a little closer to "average" detector one might use for enrichment measurements.
Clamped AD a little better for self and external attenuation - and try to keep from starting with a value of 0.0.
Reduce number of gammas from outside the ROI that get evaluated, aby a little better estimation.
If a physical model, and using Hoerl, now do a pre-fit without the Hoerl, to try and get everything in close to good range before the fit with the Hoerl - this is experimental, but seems to help for some problems.
Removed using a loss function - for a couple problems it seemed to be doing more harm than good.
Improves some spectra - not all.