Cleanup the LLM config process.
Make reading in the LLM/MCP config happen during program startup (specifically in `InterSpecServer::startWebServer(...)`), as well as starting the MCP server, if it is enabled.
Set enabling the LLM Assistant and MCP server to be seperate config settings.
Cleaned up reading in and writing the config settings, and started versioning.
Only showing option to show the LLM Assistant if configuration file is present and LLM API is enabled.