Update D3SpectrumDisplayDiv to better match how the underlying JS and CSS handle color setting.
Made setting the spectrum colors according to the color theme a application wide function (since it used global CSS variables anyway), and changed the color setting functions to be overrides of the color theme.  Movedhandling the global CSS rules to InterSpecApp, since we will probably also use it for other elements in the future.