Fixup some issues with background/secondary peaks.
The D3SpectrumDisplayDiv was displaying the primary spectrum instance background peaks for any instance of D3SpectrumDisplayDiv.
Moved to having the PeakModel also track background/secondary peaks, but in a very limited way - you can only set all the peaks for background/secondary - which triggers a single signal that these peaks were changed - not the granularity that the foreground offers.
So now D3SpectrumDisplayDiv uses the PeakModel to track and get the background/secondary peaks.