Fix issue with Act/Shield fit of trace-sources, not through the GUI
When trace sources were fit from N42 file definition, the total activity of the source was being used for the trace-source type activity (so potentially for the per cm3 or per m2 or per gram values) - this was not happening through the GUI (and it didn't really matter if you were fitting the activity for N42 definitions - only if activity was fixed). The fix is to use the trace source definitions to define the Minuit parameters - but some future work on checking, or preferably eliminating this over-specification that can be inconsistent, would be good.