For Act/Shielding fit, add the inja templates and options for showing for the calculation log.
This required re-factoring the <iframe> display stuff used to show summary results for batch analysis, into the InjaLogDialog class.
Seems to work pretty decent.
Left the old act/shield report logging in for the moment, but it will remove it at some point because it is a huge mess, and I think the new inja-template-based mecahnism covers everything (but this needs double checking) and more.