Fix user reported bug of batch peak fitting results not including some clearly fittable peaks it should have.
Some easily/clearly fittable peaks could be left out of batch peak fitting, and although not encountered that I know of, the same issues could cause user refit of peaks to fail, or potentually other places.
The cause was failure of the fit for peaks that already have essentually perfect values - my guess is the trust region was just to wildy un-matched for the current tollerances.
The problem can be fixed by:
- Setting `Ceres::Options::max_num_consecutive_invalid_steps` from its default value of 5, to somethign like 10.
- Just not re-doing the batch-fits (which was a holdover of the old Minuit2 based peak fitting)
- Or switching to numerical differentiation for the peak fit, instead of automatic differentiation

This commit implements all three fixes for robustness, even though any one of the fixes seemed to work well.

This commit also brings over the `chi2_significance_test(...)` function from the 'feature/PeakFitImprove' branch since I found a case of the old function not working well.