Work on isotopic and activity fit agents, and MDA tool-call
Change activity fit to either use a "activity_fit" tool call that uses internal state, or to use "activity_fit_one_off", for what was before custom or point source.  Previously the "custom" mode was really not using all the configuration that it claimed, so implemented hopefully using all the configuration options purported is available.
Split out implementation of activity and isotopics agents to their own source code and unit-test files.
Improved the unit tests for the agents, which caught a number of issues.
Fixed isotopics state-machine not working correctly, and added some unit tests for it (the state machine definition for the isotopics agent still needs more work).
Changed so isotopics state will also be propagated to the RelActAuto GUI, if open.
Added option to specify nuclide and distance to MDA tool call, and hopefully fixed up a few other shortcomings of it.
Various other small improvements/fixes.
AI was used to help with some of this commit.