Improve behaviour of nuclide name enter on "Reference Photopeak" tab.
May be solution for issue #19 - but not checked on Linux/Windows yet.
Fixed if you type a nuclide in fast and hit enter, the first item in suggested list would be entered (before it was updated from the C++), instead of what you typed.
Fix erroneous error message when clicking on a suggestion from the list (the error message would say what you had typed so far was an invalid nuclide)
Hopefully fixed special character handling - like the "(" not entering sometimes when trying to add a reaction.
Fix selection not being visible when you use arrow to go down past visible list (when scrollbar is present).
Improve handling width of suggestion popup - hopefully.